♻️ Refactor: redux store, reducers 분리
