+++ src/components/Sidebar/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
-import { setIsFoldSideBar } from "../../store";
+import { setIsFoldSideBar } from "../../redux/reducers/isFoldSideBar";
 import SideBarTool from "../SidebarTool";
 
 const SideBar = () => {
