+++ src/components/SidebarTool/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
-import { setSelectedTool } from "../../store";
+import { setSelectedTool } from "../../redux/reducers/selectedTool";
 import capitalizeFirstLetter from "../../utils/capitalizeFirstLetter";
 
 const SideBarTool = ({ tool }) => {
