+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import App from "./App";
-import store from "./store";
+import store from "./redux/configureStore";
 import GlobalStyle from "./style/GlobalStyle";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
