+++ src/redux/reducers/selectedTool.js
@@ -1,0 +1,15 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const selectedTool = createSlice({
+  name: "selectedTool",
+  initialState: { selectedTool: "drawing" },
+  reducers: {
+    setSelectedTool(state, action) {
+      state.selectedTool = action.payload;
+    },
+  },
+});
+
+export const { setSelectedTool } = selectedTool.actions;
+
+export default selectedTool.reducer;
